% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{canonicalize_populations}
\alias{canonicalize_populations}
\title{Check observation meta data object}
\usage{
canonicalize_populations(
  populations,
  observations,
  locations,
  max_cohort,
  max_age,
  max_dose = 2L
)
}
\arguments{
\item{populations}{a \verb{[data.frame()]}, the the observation meta data, with
columns
\itemize{
\item \code{obs_id}, any type; the observation the row concerns (i.e. id shared with
an observations data object)
\item `loc_id``, any type; the location the row concerns (i.e. id shared with a
locations data object)
\item \code{dose}, a non-zero, positive integer (1, 2, ...); what dose row concerns
\item \code{cohort}, a positive integer; the cohort at the location row concerns
\item \code{age}, a positive integer; the age of that cohort row concerns
\item \code{weight}, a numeric, (0, 1); the relative contribution of this row to an
observation
Note that multiple rows may concern the same observation, meaning that the
populations from different cohorts, locations, and ages may be pooled in an
observation
}}

\item{observations}{a \verb{[data.frame()]}, the observed data, with at least
three columns:
\itemize{
\item an \code{obs_id} column; any type, as long as unique, non-NA
\item a `positive`` column; non-negative integers, the observed number of
vaccinated individuals
\item a \code{sample_n} column; positive integers, the number of individuals
sampled, must be greater than or equal to "positive"
\item optionally, a \code{censored} column; numeric, NA (uncensored) or 1
(right-censored); if not present, will be assumed NA
}}

\item{locations}{a \verb{[data.frame()]}, with columns \code{loc_id} and \code{parent_id},
of the same type. See Details for restrictions.}

\item{max_cohort}{if present, what is the maximum cohort that should be
present?}

\item{max_age}{if present, what is the maximum age that should be present?}
}
\value{
a canonical populations object, mirroring the input \code{populations},
with the following updates:
\itemize{
\item \code{obs_c_id}, the observation id the row concerns, canonicalized to match
the canonical observation ids
\item \code{loc_c_id}, the location id the row concerns, canonicalized to match
\item reordered to \code{obs_c_id} order
}
}
\description{
Check observation meta data object
}
\details{
This method validates the meta-data associated with the observations, as well
as converting that meta-data to use the canonical id formats.

Regarding "cohorts" and "ages": these are counted from 1, by 1 "unit". You
can imagine the units are whatever resolution is appropriate for your data:
months, quarters, years, etc. As long as these are used consistently,
estimation will work, and take on the unit meaning you used for input.
}
