% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{check_locations}
\alias{check_locations}
\title{Check location data}
\usage{
check_locations(location_map)
}
\arguments{
\item{locations}{a \code{data.frame}, with integer columns \code{id} and \code{parent_id}.
See Details for restrictions.}
}
\value{
a \code{data.table}, with the canonical location structure, and a
\code{layer} column
}
\description{
Check location data
}
\details{
\code{\link[=imuGAP]{imuGAP()}} works on a hierarchical model of locations, so needs to be
told what that structure is. This method checks that the location structure
is valid, and returns a canonical version including the layer size.

\code{id} may not have an duplicates. If there is a row with \code{id == 1}, it must
have \code{parent_id == NA}. \code{max(id)} must equal the number of rows, give or
take the presence of the root \code{id == 1} - put another way, \code{id} must be
\code{seq_len(dim(locations)[1])} (again, give or take 1 for the root node).
After ordering, \code{id} and \code{parent_id} must be \code{all(diff(id) == 1)} and
\code{all(diff(parent_id) \%in\% c(0, 1))}.

Intended future capability: be able to extract a location structure that
satisfies the requirements. In particular:
\itemize{
\item allow a root other than 1; would be defined by a \code{parent_id} == \code{NA}
\item accept arbitrary column types for \code{id}, including a non-covering set of
integers, and \code{parent_id}; the return object will include an appropriate
remapping to be applied to weighting data
\item accept arbitrary layer depths
}
}
