% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{check_obs_population}
\alias{check_obs_population}
\title{Validate \code{obs_population} object for \code{imuGAP}}
\usage{
check_obs_population(
  obs_population,
  observations,
  locations,
  max_cohort,
  max_age
)

check_obs_population(
  obs_population,
  observations,
  locations,
  max_cohort,
  max_age
)
}
\arguments{
\item{obs_population}{a data.frame, the the observation weighting object, with columns
\itemize{
\item obs_id, the observation id the row concerns
\item dose, which dose the row concerns
\item location, the location the row concerns
\item cohort, the cohort at that location the row concerns
\item age, the age of that cohort the row concerns
\item weight, the relative contribution of this row to an observation
Note that multiple rows may concern the same observation, meaning that the populations
from different cohorts, locations, and ages may be pooled in an observation
}}

\item{observations}{a \code{data.frame} with:
\itemize{
\item nonnegative integer \code{positive} column
\item positive integer \code{sample_n} column, which must be greater than or equal
to the \code{positive} column
\item optionally, an \code{id} column, 1:size(observations)
}}

\item{locations}{a \code{data.frame}, with integer columns \code{id} and \code{parent_id}.
See Details for restrictions.}

\item{max_cohort}{if present, what is the maximum cohort that should be present?}

\item{max_age}{if present, what is the maximum age that should be present?}

\item{max_obs_id}{optional positive integer, the maximum valid observation id}

\item{max_location}{optional positive integer, the maximum valid location id}

\item{subset}{logical; if \code{FALSE} (default), observations must be complete.
If \code{TRUE}, \code{max_obs_id} must be provided.}
}
\value{
if \code{x} is valid, returns the canonical \code{obs_population} object
}
\description{
Validate \code{obs_population} object for \code{imuGAP}

Check observation meta data object
}
\details{
A valid \code{obs_population} object must:
\itemize{
\item be convertible to a \code{data.table} via \code{\link[data.table:setDT]{data.table::setDT()}}
\item contain the columns "obs_id", "location", "cohort", "age", "dose", and
"weight"
\item obs_id, integers >= 1; if \code{max_obs_id} is provided, <= \code{max_obs_id};
if \code{subset} is \code{FALSE}, all integers 1:\code{max_obs_id} must be present
\item location must be a positive integer less than or equal to \code{max_location}.
however, not all locations need to be represented in \code{wts}
\item cohort must be a positive integer
\item age must be a positive integer
\item dose must be 1 or 2
\item weight must be a positive numeric value; weights must sum to 1 by obs_id
}
}
\section{Errors}{

If \code{x} is not valid, raises an error describing the issue.
}

