% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{check_observations}
\alias{check_observations}
\title{Check observations objects}
\usage{
check_observations(observations)
}
\arguments{
\item{observations}{a \code{data.frame} with:
\itemize{
\item nonnegative integer \code{positive} column
\item positive integer \code{sample_n} column, which must be greater than or equal
to the \code{positive} column
\item optionally, an \code{id} column, 1:size(observations)
}}
}
\value{
the observation object, possibly changed in place to cast columns
to integers and add the \code{id} column if not original present. Will error
if the column requirements are not satisfied.
}
\description{
Check observations objects
}
