% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgap.R
\name{imgap}
\alias{imgap}
\title{Immunity: Geographic & Age-based Projection, \code{imgap}}
\usage{
imgap(
  x,
  y,
  model = c("ar_cnty", "ar_schl", "cnty_noisy", "cnty_schl_static"),
  ...
)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numeric vector of output values.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
This function estimates current coverage, by age and location.
}
