% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP-package.R
\docType{package}
\name{imuGAP-package}
\alias{imuGAP-package}
\alias{imuGAP}
\title{The 'imuGAP' package.}
\description{
A package for estimating measles vaccine coverage
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan.
R package version NA. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://accidda.github.io/imuGAP/}
  \item \url{https://github.com/ACCIDDA/imuGAP}
  \item Report bugs at \url{https://github.com/ACCIDDA/imuGAP/issues}
}

}
\author{
\strong{Maintainer}: Carl Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID})

Authors:
\itemize{
  \item Claire Perrin Smith (\href{https://orcid.org/0000-0003-1069-9460}{ORCID})
}

Other contributors:
\itemize{
  \item Kelly Zhen \email{zhen717605@gmail.com} [contributor]
  \item Weston Voglesonger \email{westonvogle@gmail.com} [contributor]
  \item Joshua Chen \email{joshuazchen1@gmail.com} [contributor]
  \item Minjae Kung \email{minjaekung@gmail.com} [contributor]
}

}
