% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{imuGAP}
\alias{imuGAP}
\title{Immunity: Geographic & Age-based Projection, \code{imuGAP}}
\usage{
imuGAP(
  observations,
  populations,
  locations,
  imugap_opts = imugap_options(),
  stan_opts = stan_options()
)
}
\arguments{
\item{observations}{a \verb{[data.frame()]}, the observed data, with at least
three columns:
\itemize{
\item an \code{obs_id} column; any type, as long as unique, non-NA
\item a `positive`` column; non-negative integers, the observed number of
vaccinated individuals
\item a \code{sample_n} column; positive integers, the number of individuals
sampled, must be greater than or equal to "positive"
\item optionally, a \code{censored} column; numeric, NA (uncensored) or 1
(right-censored); if not present, will be assumed NA
}}

\item{populations}{a \verb{[data.frame()]}, the the observation meta data, with
columns
\itemize{
\item \code{obs_id}, any type; the observation the row concerns (i.e. id shared with
an observations data object)
\item `loc_id``, any type; the location the row concerns (i.e. id shared with a
locations data object)
\item \code{dose}, a non-zero, positive integer (1, 2, ...); what dose row concerns
\item \code{cohort}, a positive integer; the cohort at the location row concerns
\item \code{age}, a positive integer; the age of that cohort row concerns
\item \code{weight}, a numeric, (0, 1); the relative contribution of this row to an
observation
Note that multiple rows may concern the same observation, meaning that the
populations from different cohorts, locations, and ages may be pooled in an
observation
}}

\item{locations}{a \verb{[data.frame()]}, with columns \code{loc_id} and \code{parent_id},
of the same type. See Details for restrictions.}

\item{imugap_opts}{options for the \code{imuGAP} model}

\item{stan_opts}{passed to \verb{[rstan::sampling()]} (e.g. \code{iter}, \code{chains}).}
}
\value{
An object of class \code{stanfit} returned by \verb{[rstan::sampling()]}
}
\description{
This a sampler interface to convert user-friendly data into the necessary
format to feed the immunity estimation model.
}
\examples{
\dontrun{
data("locations_sim"); data("observations_sim"); data("populations_sim")
st_opts <- stan_options(chains = 2, iter = 500)
imuGAP(
  observations_sim, populations_sim, locations_sim,
  stan_opts = st_opts()
)
}

}
