% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{imugap_options}
\alias{imugap_options}
\title{imuGAP Model Options}
\usage{
imugap_options(df = 5L, dose_schedule = c(1, 4), object = c("default"))
}
\arguments{
\item{df}{degrees of freedom to use in bspline}

\item{dose_schedule}{an integer vector, the ages at which dose(s) \code{n} are
scheduled, with vector indices and doses matching}

\item{object}{which stan model object to use; currently only "default" is
supported}
}
\value{
a list of imuGAP model options
}
\description{
This function encapsulates option passing for imuGAP settings.
}
\examples{
imugap_options() # defaults
imugap_options(dose_schedule = c(1, 3)) # change dose schedule model

}
