% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observations_sim}
\alias{observations_sim}
\title{Simulated observations}
\format{
A \link[data.table:data.table]{data.table::data.table} with columns:
\describe{
\item{obs_id}{Observation identifier (integer)}
\item{positive}{Number of vaccinated individuals}
\item{sample_n}{Total sample size}
}
}
\usage{
observations_sim
}
\description{
Simulated vaccination observation data matching \link{locations_sim}.
}
\keyword{datasets}
