% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{populations_sim}
\alias{populations_sim}
\title{Simulated observation populations}
\format{
A \link[data.table:data.table]{data.table::data.table} with columns:
\describe{
\item{obs_id}{Observation identifier matching \link{observations_sim}}
\item{loc_id}{Location identifier matching \link{locations_sim}}
\item{cohort}{Cohort index (positive integer)}
\item{age}{Age index (positive integer)}
\item{dose}{Dose number (1 or 2)}
\item{weight}{Relative contribution to the observation (0, 1]}
}
}
\usage{
populations_sim
}
\description{
Observation-to-population mapping for \link{observations_sim}, linking each
observation to location, cohort, age, dose, and weight.
}
\keyword{datasets}
