% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{predict.imugap_fit}
\alias{predict.imugap_fit}
\title{Predict coverage probabilities}
\usage{
\method{predict}{imugap_fit}(object, populations, ...)
}
\arguments{
\item{object}{an \code{imugap_fit} object returned by \code{sampling()}}

\item{populations}{a \verb{[data.frame()]} of target populations to predict for}

\item{...}{additional arguments passed to other methods.}
}
\value{
A \code{data.table} with columns \code{sample_id}, \code{obs_id}, and \code{p_obs} containing
the predicted coverage probabilities for each posterior draw and target observation.
}
\description{
Uses the output of \code{\link{sampling}} and a target \code{populations} grid to generate
predicted coverage probabilities.
}
