% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{sampling}
\alias{sampling}
\title{Immunity: Geographic & Age-based Projection, \code{imuGAP}}
\usage{
sampling(
  observations,
  populations,
  locations,
  imugap_opts = imugap_options(),
  stan_opts = stan_options()
)
}
\arguments{
\item{observations}{a \verb{[data.frame()]}, the observed data, with at least
three columns:
\itemize{
\item an \code{obs_id} column; any type, as long as unique, non-NA
\item a \code{positive} column; non-negative integers, the observed number of
vaccinated individuals
\item a \code{sample_n} column; positive integers, the number of individuals
sampled, must be greater than or equal to "positive"
\item optionally, a \code{censored} column; numeric, NA (uncensored) or 1
(right-censored); if not present, will be assumed NA
}}

\item{populations}{a \verb{[data.frame()]}, the observation meta data, with
columns
\itemize{
\item \code{obs_id}, any type; the observation the row concerns (i.e. id shared with
an observations data object)
\item \code{loc_id}, any type; the location the row concerns (i.e. id shared with a
locations data object)
\item \code{dose}, a non-zero, positive integer (1, 2, ...); what dose row concerns
\item \code{cohort}, a positive integer; the cohort at the location row concerns
\item \code{age}, a positive integer; the age of that cohort row concerns
\item \code{weight}, a numeric, (0, 1); the relative contribution of this row to an
observation. Optional if each population row has a unique \code{obs_id}.
}}

\item{locations}{a \verb{[data.frame()]}, with columns \code{loc_id} and \code{parent_id},
of the same type. See Details for restrictions.}

\item{imugap_opts}{options for the \code{imuGAP} model}

\item{stan_opts}{passed to \code{rstan::sampling} (e.g. \code{iter}, \code{chains}).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
This a sampler interface to convert user-friendly data into the necessary
format to feed the immunity estimation model.
}
\examples{
\dontrun{
data("locations_sim"); data("observations_sim"); data("populations_sim")
st_opts <- stan_options(chains = 2, iter = 500)
imuGAP(
  observations_sim, populations_sim, locations_sim,
  stan_opts = st_opts
)
}

}
