% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{stan_options}
\alias{stan_options}
\title{Stan Sampler Options}
\usage{
stan_options(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}}
  \describe{
    \item{\code{object}}{An object of class \code{\linkS4class{stanmodel}}.}
  }}
}
\value{
a list of arguments matching \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} inputs
}
\description{
This function encapsulates option passing to the stan sampler, with the
exception of the model object, which is passed in \code{imugap_options}.
}
