% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{extract_imugap}
\alias{extract_imugap}
\title{Custom imuGAP fit extraction}
\usage{
extract_imugap(fit, pars = c("logit_phi_state"), ...)
}
\arguments{
\item{fit}{an \code{imugap_fit} object returned by \code{sampling()}}

\item{pars}{character vector; parameters to extract.}

\item{...}{additional arguments passed to \verb{[rstan::extract()]}.}
}
\value{
a list, as returned by \code{rstan::extract()}
}
\description{
Thin wrapper around \code{rstan::extract} to extract typical imuGAP parameters.
}
\examples{
data("fit_sim")
extract_imugap(fit_sim)
extract_imugap(fit_sim, pars = "phi")

}
