% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP-package.R
\docType{package}
\name{imuGAP-package}
\alias{imuGAP-package}
\alias{imuGAP}
\title{The 'imuGAP' package.}
\description{
A package for estimating measles vaccine coverage
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. https://mc-stan.org
}
\author{
\strong{Maintainer}: Carl A. B. Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID})

}
