% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{imuGAP}
\alias{imuGAP}
\title{Immunity: Geographic & Age-based Projection, \code{imuGAP}}
\usage{
imuGAP(
  observations,
  populations,
  locations,
  imugap_opts = imugap_options(),
  stan_opts = stan_options()
)
}
\arguments{
\item{observations}{a \verb{[data.frame()]}, the observed data, with at least
three columns:
\itemize{
\item an "id" column; any type, as long as unique, non-NA
\item a "positive" column; non-negative integers, the observed number of
vaccinated individuals
\item a "sample_n" column; positive integers, the number of individuals
sampled, must be greater than or equal to "positive"
\item optionally, a "censored" column; numeric, NA (uncensored) or 1
(right-censored);
}}

\item{populations}{a \verb{[data.frame()]}, the the observation meta data, with
columns
\itemize{
\item "obs_id", the observation id the row concerns
\item "loc_id", the location id the row concerns
\item "dose", which dose the row concerns
\item "cohort", the cohort at that location the row concerns
\item "age", the age of that cohort the row concerns
\item "weight", the relative contribution of this row to an observation
Note that multiple rows may concern the same observation, meaning that the
populations from different cohorts, locations, and ages may be pooled in an
observation
}}

\item{locations}{a \verb{[data.frame()]}, with columns \code{id} and \code{parent_id}, of
the same type. See Details for restrictions.}

\item{imugap_opts}{options for the \code{imuGAP} model}

\item{stan_opts}{passed to \code{rstan::sampling} (e.g. \code{iter}, \code{chains}).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
This a sampler interface to convert user-friendly data into the necessary
format to feed the immunity estimation model.
}
