% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{imuGAP}
\alias{imuGAP}
\title{Immunity: Geographic & Age-based Projection, \code{imuGAP}}
\usage{
imuGAP(
  x,
  y,
  state_model = c("ar", "noisy"),
  cnty_model = c("ar", "noisy", "static"),
  schl_model = c("ar", "noisy", "static"),
  imgap_options = imgap_options(),
  stan_options = stan_options()
)
}
\arguments{
\item{x}{a}

\item{y}{Numeric vector of output values.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
This function estimates current coverage, by age and location.
}
