% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{imugap_options}
\alias{imugap_options}
\title{Options for imuGAP model}
\usage{
imugap_options(df = 5L)
}
\arguments{
\item{df}{degrees of freedom to use in bspline}
}
\value{
a list of imuGAP model options
}
\description{
This function encapsulates option passing for imuGAP settings
}
