% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP-package.R
\docType{data}
\name{locations_sim}
\alias{locations_sim}
\title{Example Location Data}
\format{
A \verb{[data.table()]} with 28 rows and 2 columns:
\itemize{
\item \code{loc_id}, a string, the location
\item \code{parent_id}, a string, the location parents
}
}
\usage{
locations_sim
}
\description{
A dataset providing example location input.
}
\keyword{datasets}
