% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP-package.R
\docType{data}
\name{observations_sim}
\alias{observations_sim}
\title{Example Observation Data}
\format{
A \verb{[data.table()]} with 698 rows and 4+ columns:
\itemize{
\item \code{obs_id}, a number, the observation id (primary key)
\item \code{positive}, a number, how many individuals had the vaccine
\item \code{sample_n}, a number, the number of individuals in the observations
\item \code{censored}, a number, 1 if \code{positive} is right censored and
\code{NA} if uncensored
\item ... assorted other columns that are irrelevant to use as observations arg
}
}
\usage{
observations_sim
}
\description{
A dataset containing vaccine coverage observations.
}
\keyword{datasets}
