% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP-package.R
\docType{data}
\name{populations_sim}
\alias{populations_sim}
\title{Example Population Data}
\format{
A \verb{[data.table()]} with 750 rows and 6 columns:
\itemize{
\item \code{obs_id}, a number, the observation (foreign key to observations)
\item \code{loc_id}, a string, the location (foreign key to locations)
\item \code{cohort}, a number, the birth cohort
\item \code{age}, a number, the age of cohort at time of observation
\item \code{dose}, a number, which dose the observation concerns
\item \code{weight}, a number (0-1), fraction of the observation this row represents
}
}
\usage{
populations_sim
}
\description{
A dataset containing the meta-data about vaccine coverage observations.
}
\keyword{datasets}
