% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imuGAP.R
\name{stan_options}
\alias{stan_options}
\title{Options for stan sampler}
\usage{
stan_options(object = stanmodels$impute_school_coverage_process_v5, ...)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{stanmodel}}.}

\item{...}{
  Arguments passed on to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
a list of arguments matching \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} inputs
}
\description{
This function encapsulates option passing to the stan sampler.
}
